Bump versions, and depend on Class::Method::Modifiers 2.16
Class::Method::Modifiers version < 2.16 are subtly broken in
Perl 5.37.9, this in turn breaks MIME::Signature/MIME::Disclaimer.

I have created https://github.com/moose/Class-Method-Modifiers/pull/9
which should fix Class::Method::Modifiers. This patch should be
applied once that (or an equivalent fix) is merged and a new version
is released.

I have asked the owner of that repo to coordinate with this PR.

See:
https://github.com/Perl/perl5/pull/20357
https://rt.cpan.org/Ticket/Display.html?id=146848
https://github.com/Perl/perl5/issues/20885
https://github.com/fany/MIME-Signature/issues/3